New day: 2023-03-31

+++ StaticComponents/FinishedDistribution.cs
@@ -25,7 +25,7 @@ namespace FloorSimulation
         
         private int TotalTrolleysDistr;
         private int TotalTrolleysExp;
-        private float AveragePercentageFull;
+        private double AveragePercentageFull;
 
         public FinishedDistribution(Floor Floor)
         {
